+++ ag_data/formulas/library/usr/alice_formulas.py
@@ -1,0 +1,6 @@
+#
+# Alice's Formulas for example sensor types
+#
+#
+
+from ag_data import models
