+++ ag_data/formulas/library/usr/bob_formulas.py
@@ -1,0 +1,6 @@
+#
+# Bob's Formulas for example sensor types
+#
+#
+
+from ag_data import models
